requirements.txt
Django=1.10

Pillow
django-crispy-forms